Partially moved QR scanner code to FAB menu
