+++ vector/src/main/java/im/vector/app/features/userdirectory/KnownUsersFragment.kt
@@ -16,13 +16,11 @@
 
 package im.vector.app.features.userdirectory
 
-import android.app.Activity
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
 import android.widget.ScrollView
-import android.widget.Toast
 import androidx.core.view.forEach
 import androidx.core.view.isVisible
 import com.airbnb.mvrx.activityViewModel
@@ -35,29 +33,18 @@ import im.vector.app.R
 import im.vector.app.core.extensions.cleanup
 import im.vector.app.core.extensions.configureWith
 import im.vector.app.core.extensions.hideKeyboard
-import im.vector.app.core.extensions.registerStartForActivityResult
 import im.vector.app.core.extensions.setupAsSearch
 import im.vector.app.core.platform.VectorBaseFragment
 import im.vector.app.core.utils.DimensionConverter
-import im.vector.app.core.utils.PERMISSIONS_FOR_TAKING_PHOTO
-import im.vector.app.core.utils.checkPermissions
-import im.vector.app.core.utils.registerForPermissionsResult
-import im.vector.app.features.createdirect.CreateDirectRoomAction
-import im.vector.app.features.createdirect.CreateDirectRoomViewModel
 import im.vector.app.features.homeserver.HomeServerCapabilitiesViewModel
-import im.vector.app.features.qrcode.QrCodeScannerActivity
 import kotlinx.android.synthetic.main.fragment_known_users.*
 import org.matrix.android.sdk.api.session.user.model.User
-import org.matrix.android.sdk.api.session.Session
-import org.matrix.android.sdk.api.session.permalinks.PermalinkData
-import org.matrix.android.sdk.api.session.permalinks.PermalinkParser
 import javax.inject.Inject
 
 class KnownUsersFragment @Inject constructor(
         val userDirectoryViewModelFactory: UserDirectoryViewModel.Factory,
         private val knownUsersController: KnownUsersController,
         private val dimensionConverter: DimensionConverter,
-        private val session: Session,
         val homeServerCapabilitiesViewModelFactory: HomeServerCapabilitiesViewModel.Factory
 ) : VectorBaseFragment(), KnownUsersController.Callback {
 
@@ -68,7 +55,6 @@ class KnownUsersFragment @Inject constructor(
     override fun getMenuRes() = args.menuResId
 
     private val userDirViewModel: UserDirectoryViewModel by activityViewModel()
-    private val createDirectRoomViewModel: CreateDirectRoomViewModel by activityViewModel()
     private val homeServerCapabilitiesViewModel: HomeServerCapabilitiesViewModel by fragmentViewModel()
 
     private lateinit var sharedActionViewModel: UserDirectorySharedActionViewModel
@@ -82,7 +68,6 @@ class KnownUsersFragment @Inject constructor(
         setupRecyclerView()
         setupFilterView()
         setupAddByMatrixIdView()
-        setupAddByQrCodeView()
         setupAddFromPhoneBookView()
         setupCloseView()
 
@@ -119,52 +104,6 @@ class KnownUsersFragment @Inject constructor(
     private fun setupAddByMatrixIdView() {
         addByMatrixId.debouncedClicks {
             sharedActionViewModel.post(UserDirectorySharedAction.OpenUsersDirectory)
-        }
-    }
-
-    private fun setupAddByQrCodeView() {
-        val qrStartForActivityResult = registerStartForActivityResult { activityResult ->
-            if (activityResult.resultCode == Activity.RESULT_OK) {
-                val result = QrCodeScannerActivity.getResultText(activityResult.data)!!
-                val mxid = (PermalinkParser.parse(result) as? PermalinkData.UserLink)?.userId
-
-                if (mxid === null) {
-                    Toast.makeText(requireContext(), R.string.invalid_qr_code_uri, Toast.LENGTH_SHORT).show()
-                } else {
-                    val existingDm = session.getExistingDirectRoomWithUser(mxid)
-
-                    if (existingDm === null) {
-                        // The following assumes MXIDs are case insensitive
-                        if (mxid.equals(other = session.myUserId, ignoreCase = true)) {
-                            Toast.makeText(requireContext(), R.string.cannot_dm_self, Toast.LENGTH_SHORT).show()
-                        } else {
-                            // Try to get user from known users and fall back to creating a User object from MXID
-                            val qrInvitee = if (session.getUser(mxid) != null) session.getUser(mxid)!! else User(mxid, null, null)
-
-                            createDirectRoomViewModel.handle(
-                                    CreateDirectRoomAction.CreateRoomAndInviteSelectedUsers(setOf(PendingInvitee.UserPendingInvitee(qrInvitee)))
-                            )
-                        }
-                    } else {
-                        navigator.openRoom(requireContext(), existingDm.roomId, null, false)
-                        requireActivity().finish()
-                    }
-                }
-            } else {
-                Toast.makeText(requireContext(), R.string.qr_code_not_scanned, Toast.LENGTH_SHORT).show()
-            }
-        }
-        val openCameraActivityResultLauncher = registerForPermissionsResult { allGranted ->
-            if (allGranted) {
-                QrCodeScannerActivity.startForResult(requireActivity(), qrStartForActivityResult)
-            } else {
-                Toast.makeText(requireContext(), R.string.missing_permissions_error, Toast.LENGTH_SHORT).show()
-            }
-        }
-        addByQrCode.debouncedClicks {
-            if (checkPermissions(PERMISSIONS_FOR_TAKING_PHOTO, requireActivity(), openCameraActivityResultLauncher)) {
-                QrCodeScannerActivity.startForResult(requireActivity(), qrStartForActivityResult)
-            }
         }
     }
