+++ vector/src/main/res/layout/fragment_known_users.xml
@@ -139,23 +139,6 @@
             app:layout_constraintTop_toBottomOf="@id/knownUsersE2EbyDefaultDisabled" />
 
         <com.google.android.material.button.MaterialButton
-            android:id="@+id/addByQrCode"
-            style="@style/VectorButtonStyleText"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            android:layout_marginTop="8dp"
-            android:layout_marginBottom="8dp"
-            android:minHeight="@dimen/layout_touch_size"
-            android:text="@string/add_by_qr_code"
-            android:visibility="visible"
-            app:icon="@drawable/ic_plus_circle"
-            app:iconPadding="13dp"
-            app:iconTint="@color/riotx_accent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/addByMatrixId" />
-
-        <com.google.android.material.button.MaterialButton
             android:id="@+id/addFromPhoneBook"
             style="@style/VectorButtonStyleText"
             android:layout_width="0dp"
@@ -170,7 +153,7 @@
             app:iconPadding="13dp"
             app:iconTint="@color/riotx_accent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/addByQrCode" />
+            app:layout_constraintTop_toBottomOf="@id/addByMatrixId" />
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView"
